MovieAPI/requirements.txt
Django~=3.2
Pillow~=8.2.0
PyJWT~=1.7.1
asgiref~=3.3.4
django-debug-toolbar~=3.2.1
django-filter~=2.4.0
djangorestframework~=3.12.4
djangorestframework-jwt~=1.11.0
pip~=21.0.1
psycopg2~=2.8.6
pytz~=2021.1
setuptools~=56.0.0
sqlparse~=0.4.1